requirements.txt
streamlit
replicate
Pillow
requests